instances: finish migrating startup modules
