+++ src/smi_beamline/instances/__init__.py
@@ -24,23 +24,23 @@ import time
 # Each entry: (label, module_name).
 DEVICE_MODULES = [
     ("sample chamber", "smi_beamline.instances.waxschamber"),
-    ("shutters",       "smibase.shutter"),
+    ("shutters",       "smi_beamline.instances.shutter"),
     ("beamstop",       "smi_beamline.instances.beamstop"),
     ("machine/ring",   "smi_beamline.instances.machine"),
-    ("attenuators",    "smibase.attenuators"),
+    ("attenuators",    "smi_beamline.instances.attenuators"),
     ("CRLs",           "smi_beamline.instances.crls"),
-    ("manipulators",   "smibase.manipulators"),
-    ("mirrors",        "smibase.mirrors"),
+    ("manipulators",   "smi_beamline.instances.manipulators"),
+    ("mirrors",        "smi_beamline.instances.mirrors"),
     ("motors",         "smi_beamline.instances.motors"),
     ("slits",          "smi_beamline.instances.slits"),
-    ("energy",         "smibase.energy"),
+    ("energy",         "smi_beamline.instances.energy"),
     ("xbpms",          "smi_beamline.instances.xbpms"),
     ("ioLogik",        "smi_beamline.instances.ioLogik"),
     ("electrometers",  "smi_beamline.instances.electrometers"),
     ("amptek",         "smi_beamline.instances.amptek"),
-    ("pilatus",        "smibase.pilatus"),
-    ("prosilica",      "smibase.prosilica"),
-    ("beam modes",     "smibase.beam"),
+    ("pilatus",        "smi_beamline.instances.pilatus"),
+    ("prosilica",      "smi_beamline.instances.prosilica"),
+    ("beam modes",     "smi_beamline.instances.beam"),
     ("alignment",      "smi_beamline.plans.alignment"),
     ("config",         "smi_beamline.plans.config"),
     ("scan naming",    "smi_beamline.plans.scan_naming"),
@@ -48,7 +48,7 @@ DEVICE_MODULES = [
     ("blade coater",   "smi_beamline.instances.bladecoater"),
     ("humidity cell",  "smi_beamline.plans.humidity_cell"),
     ("linkam",         "smi_beamline.instances.linkam"),
-    ("suspenders",     "smibase.suspenders"),
+    ("suspenders",     "smi_beamline.instances.suspenders"),
     ("utils",          "smi_beamline.plans.utils"),
 ]
