+++ src/smi_beamline/instances/beam.py
@@ -1,0 +1,700 @@
+
+
+
+import numpy as np
+import pandas as pd
+import os
+import bluesky.plan_stubs as bps
+from smi_beamline.instances.pilatus import pil2m_pos
+from smi_beamline.instances.beamstop import saxs_bs
+from smi_beamline.devices import _context
+RE = _context.get_re()
+from smi_beamline.instances.attenuators import (att1_1, att1_2, att1_3, att1_4, att1_5, att1_6, att1_7, att1_8, att1_9, att1_10, att1_11, att1_12,
+                                 att2_1, att2_2, att2_3, att2_4, att2_5, att2_6, att2_7, att2_8, att2_9, att2_10, att2_11, att2_12,
+                                 attenuation)
+from smi_beamline.instances.energy import energy
+from smi_beamline.instances.pilatus import waxs, pil2M, pil900KW
+from smi_beamline.instances.beamstop import saxs_bs
+from smi_beamline.instances.crls import crl
+from smi_beamline.instances.waxschamber import get_chamber_pressure, chamber_pressure
+
+from ophyd import Signal, Component, Device
+
+# devices to store metadata about detectors
+class SMI_WAXS_detector(Device):
+    prefix = "detector_waxs_"
+    pixel_size = Component(Signal, value=0.172, name=prefix + "pixel_size", kind="hinted")
+    x0_pix = Component(Signal, value=97, name=prefix + "x0_pix", kind="hinted")
+    y0_pix = Component(Signal, value=1386, name=prefix + "y0_pix", kind="hinted")
+    sdd = Component(Signal, value=274.9, name=prefix + "sdd", kind="hinted")
+
+
+class SMI_SAXS_detector(Device):
+    prefix = "detector_saxs_"
+    pixel_size = Component(Signal, value=0.172, name=prefix + "pixel_size", kind="hinted")
+    bs_kind = Component(Signal, value="rod", name=prefix + "bs_kind", kind="hinted")
+    xbs_mask = Component(Signal, value=0, name=prefix + "xbs_mask", kind="hinted")
+    ybs_mask = Component(Signal, value=0, name=prefix + "ybs_mask", kind="hinted")
+    x0_pix = Component(Signal, value=0, name=prefix + "y0_pix", kind="hinted")
+    y0_pix = Component(Signal, value=0, name=prefix + "y0_pix", kind="hinted")
+    sdd = Component(Signal, value=8300, name=prefix + "sdd", kind="hinted")
+
+
+class SMIBeam(object):
+    """
+    This class represents the 'beam' at the beamline. This collects together aspects
+    of querying or changing beam properties, including the energy (or wavelength),
+    the beam intensity (or measuring flux), and so forth.
+    """
+
+    def __init__(self):
+
+        self._SHUTTER_CLOSED_VOLTAGE = 7
+        self.hc_over_e = 1.23984197e-6  # eV*m
+        self.D_Si111 = 3.1293  # Angstroms
+        self.dcm = energy
+
+    def _foilState(self):
+        current_state = []
+        for foil_num in [att1_1, att1_2, att1_3, att1_4, att1_5, att1_6, att1_7, att1_8, att1_9, att1_10, att1_11, att1_12,
+                         att2_1, att2_2, att2_3, att2_4, att2_5, att2_6, att2_7, att2_8, att2_9, att2_10, att2_11, att2_12]:
+            foil = yield from bps.read(foil_num)
+            if foil["{}_status".format(foil_num.name)]["value"] == "Open":
+                current_state.append(foil_num)
+        return current_state
+
+    def _determineFoils(self):
+        """Return the test alignment attenuation factor for the new attenuation object."""
+        return 1e6
+
+        """
+        Previous fixed individual-foil table, kept inactive while testing the
+        energy-aware attenuation object.
+
+        if self.dcm.energy.position < 2000:
+            target_state = [att1_12]
+        elif 2000 < self.dcm.energy.position < 2300:
+            target_state = [att2_5, att2_6]
+        elif 2300 < self.dcm.energy.position < 2700:
+            target_state = [att2_10, att2_11, att2_12]
+        elif 2700 < self.dcm.energy.position < 3000:
+            target_state = [att2_10, att2_12]
+        elif 3000 < self.dcm.energy.position < 3200:
+            target_state = [att2_9, att2_10, att2_11]
+        elif 3200 < self.dcm.energy.position <= 3500:
+            target_state = [att2_9, att2_10, att2_11, att2_12]
+        elif 3500 < self.dcm.energy.position <= 3750:
+            target_state = [att2_5, att2_9]
+        elif 3750 < self.dcm.energy.position <= 3900:
+            target_state = [att2_5, att2_9]
+        elif 3900 < self.dcm.energy.position < 4200:
+            target_state = [att2_5, att2_9, att2_10]
+        elif 4200 < self.dcm.energy.position < 4500:
+            target_state = [att2_6, att2_7]
+        elif 4500 < self.dcm.energy.position < 5500:
+            target_state = [att2_5, att2_6]
+        elif 5500 < self.dcm.energy.position < 7000:
+            target_state = [att1_12]
+        elif 7000 < self.dcm.energy.position < 7500:
+            target_state = [att2_5, att2_6, att2_7]
+        elif 7500 < self.dcm.energy.position < 8400:
+            target_state = [att2_1, att2_3]
+        elif 8400 < self.dcm.energy.position < 8800:
+            target_state = [att2_2, att2_3]
+        elif 8800 < self.dcm.energy.position < 9700:
+            #target_state = [att2_1, att2_2, att2_3]
+            target_state = [att2_3]
+        elif 9700 < self.dcm.energy.position < 10500:
+            #target_state = [att2_4] # changing for lower flux
+            target_state = [att2_3]
+        elif 10500 < self.dcm.energy.position < 11000:
+            target_state = [att2_1, att2_4]
+        elif 11000 < self.dcm.energy.position < 11500:
+            target_state = [att1_10, att1_11]
+        elif 11500 < self.dcm.energy.position < 13000:
+            target_state = [att1_9, att1_10, att1_11]
+        elif 13000 < self.dcm.energy.position < 14000:
+            target_state = [att1_10, att1_12]
+        elif 14000 < self.dcm.energy.position < 14700:
+            target_state = [att1_6, att1_7]
+        elif 14700 < self.dcm.energy.position < 16100:
+            target_state = [att1_5, att1_6, att1_7]
+            #For microfiocusing
+            # target_state = [att1_7, att1_6]
+        elif 16100 < self.dcm.energy.position < 17500:
+            target_state = [att1_8]
+        elif 17500 < self.dcm.energy.position < 18500:
+            target_state = [att1_1, att1_3]
+        elif 18500 < self.dcm.energy.position < 20000:
+            target_state = [att1_2, att1_3]
+        elif 20000 < self.dcm.energy.position < 23001: #added 1 here -Tom Chaney
+            target_state = [att1_1, att1_2, att1_3]
+        return target_state
+        """
+
+    def insertFoils(self, num_foils):
+        if num_foils == "Measurement":
+            yield from bps.mv(attenuation, 1)
+        else:
+            yield from bps.mv(attenuation, self._determineFoils())
+        return
+
+        """
+        Previous individual-foil actuation path, kept inactive while testing the
+        energy-aware attenuation object.
+
+        if num_foils == "Measurement":
+            target_state = []
+        else:
+            target_state = self._determineFoils()
+        current_state = yield from self._foilState()
+
+        # First insert foils
+        for foil in target_state:
+            if foil not in current_state:
+                yield from self._actuateFoil(foil, "Insert")
+
+        # Then remove foils not needed
+        for foil in current_state:
+            if foil not in target_state:
+                yield from self._actuateFoil(foil, "Retract")
+
+        # Double check that it worked
+        # This does not work
+        current_state = yield from self._foilState()
+        if current_state != target_state:
+            print("WARNING: Foils did not actuate correctly")
+            print("current state: {}".format(current_state))
+            print("target state: {}".format(target_state))
+        """
+
+    def _actuateFoil(self, foil, state, wait_time=2.0, max_retries=5):
+        itry = 0
+        foil_st = yield from bps.read(foil)
+
+        while itry < max_retries and (( foil_st["{}_status".format(foil.name)]["value"] == "Not Open" and state == "Insert")
+                                      or (foil_st["{}_status".format(foil.name)]["value"] == "Open" and state == "Retract")):
+            if state == "Retract":
+                yield from bps.mv(foil.close_cmd, 1)
+            elif state == "Insert":
+                yield from bps.mv(foil.open_cmd, 1)
+            itry += 1
+            yield from bps.sleep(wait_time)
+            foil_st = yield from bps.read(foil)
+
+    def calc_bswaxs_posy(self):
+        # Move the waxs beamstop up for safety. If tender, covers up to 1deg ai, if hard, covers up to 0.4deg ai
+        if self.dcm.energy.position < 6000:
+            yield from bps.mv(waxs.bs_y, -0.05)
+        else:
+            # it is -7 for transmission, around -2 to -3 for reflection
+            yield from bps.mv(waxs.bs_y, -0.05)
+
+# End class SMIBeam(object)
+########################################
+
+beam = SMIBeam()
+
+
+class Beamline(object):
+    """
+    Generic class that encapsulates different aspects of the beamline.
+    The intention for this object is to have methods that activate various 'standard'
+    protocols or sequences of actions.
+    """
+
+    def __init__(self, **kwargs):
+        self.md = {}
+        self.current_mode = "undefined"
+
+    def update_md(self, **md):
+        """
+        Returns a dictionary of the updated metadata.
+        """
+        self.md.copy()
+
+
+class SMI_Beamline(Beamline): # used in alignment
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        # Metadata
+        self.md = kwargs
+        self.SAXS = SMI_SAXS_Det()
+        self.WAXS = SMI_WAXS_Det() # for bdm using 900KW
+
+        self.attenuators_state()
+        # self.crl_state()
+        self.pressure_measurments()
+
+        self._pil2m_gap_restore_y = None
+        self._pil2m_gap_y_pixel_shift = 0.0
+
+        self.update_md()
+
+    def modeAlignment(self, technique="gisaxs"):
+        """
+        Set the beamline for alignement: move the beamstop out and insert the attenuators
+        if technique is 'gisaxs', the reflected roi will be calculate for horizontal congif
+        if technique is 'xrr', the reflected roi will be calculate for vertical congif
+        """
+        # Put in attenuators
+        yield from SMIBeam().insertFoils("Alignement")
+
+        # move beamstop (whichever is active) out of direct beam by 5 mm
+        yield from pil2M.remove_beamstop()
+
+        # Move the waxs beamstop up for safety. If tender, covers up to 1deg ai, if hard, covers up to 0.4deg ai
+        yield from SMIBeam().calc_bswaxs_posy()
+
+        yield from self.setReflectedBeamROI(technique=technique)
+        yield from self.setDirectBeamROI()
+
+        if technique == "gisaxs":
+            # Move the waxs detector out of the way
+            if waxs.arc.position < 15:
+                yield from bps.mv(waxs.arc, 15)
+        
+        elif technique == "giwaxs": # for bdm
+            print('giwaxs mode: move waxs to 0 deg!')
+            yield from bps.mv(waxs.arc, 0)
+
+    def modeMeasurement(self):
+        """
+        Set the beamline for measurments: bring the beamstop in and
+        remove the attenuator
+        """
+        # Move beamstop back into position
+        yield from pil2M.restore_beamstop()
+
+        # Remove attenuators
+        yield from SMIBeam().insertFoils("Measurement")
+
+    def setDirectBeamROI(self, size=[48, 12], technique="gisaxs"):
+        """
+        Update the ROI (pil2m.roi1) for the direct beam on the SAXS detector.
+        size: tuple argument: size in pixels) of the ROI [width, height]).
+        """
+        #Reading the current position of 2M to adjust ROI position
+        self.SAXS.getPositions()
+
+        # These positions are updated based on current detector position
+        x0 = self.SAXS.direct_beam[0]
+        y0 = self.SAXS.direct_beam[1]
+
+        if technique == "gisaxs":
+            # Define the direct beam ROI on the pilatus 2M detector
+            yield from bps.mv(pil2M.roi1.min_xyz.min_x, int(x0 - size[0] / 2))
+            yield from bps.mv(pil2M.roi1.size.x, int(size[0]))
+            yield from bps.mv(pil2M.roi1.min_xyz.min_y, int(y0 - size[1] / 2))
+            yield from bps.mv(pil2M.roi1.size.y, int(size[1]))
+
+        elif technique == "xrr":
+            # Define the direct beam ROI on the pilatus 2M detector
+            yield from bps.mv(pil2M.roi1.min_xyz.min_x, int(x0 - size[1] / 2))
+            yield from bps.mv(pil2M.roi1.size.x, int(size[1]))
+            yield from bps.mv(pil2M.roi1.min_xyz.min_y, int(y0 - size[0] / 2))
+            yield from bps.mv(pil2M.roi1.size.y, int(size[0]))
+        
+        elif technique == "giwaxs":  # for bdm using 900KW
+
+            #Reading the current position of 2M to adjust ROI position
+            self.WAXS.getPositions()
+
+            # These positions are updated based on current detector position
+            x0 = self.WAXS.direct_beam[0]
+            y0 = self.WAXS.direct_beam[1]
+
+            yield from bps.mv(pil900KW.roi1.min_xyz.min_x, int(x0 - size[1] / 2))
+            yield from bps.mv(pil900KW.roi1.size.x, int(size[1]))
+            yield from bps.mv(pil900KW.roi1.min_xyz.min_y, int(y0 - size[0] / 2))
+            yield from bps.mv(pil900KW.roi1.size.y, int(size[0]))
+
+
+    @staticmethod
+    def _roi_overlaps_gap(roi_min, roi_size, gaps):
+        roi_max = roi_min + roi_size
+        return any(roi_min < gap_max and roi_max > gap_min for gap_min, gap_max in gaps)
+
+    @classmethod
+    def _roi_gap_clear_shift(cls, roi_min, roi_size, gaps, margin=4):
+        """Pixel shift needed to move an ROI fully out of the detector gap."""
+        roi_max = roi_min + roi_size
+        candidates = []
+
+        for gap_min, gap_max in gaps:
+            if roi_min < gap_max and roi_max > gap_min:
+                candidates.append(gap_min - margin - roi_max)
+                candidates.append(gap_max + margin - roi_min)
+
+        for shift in sorted(candidates, key=abs):
+            if not cls._roi_overlaps_gap(roi_min + shift, roi_size, gaps):
+                return shift
+
+        return 0
+
+    @staticmethod
+    def _pil2m_motor_y_delta_for_pixel_shift(pixel_shift, pixel_size):
+        """Convert desired detector-pixel Y shift to a pil2M Y-motor move."""
+        beam_y_mm = pil2M.beam_center_y_mm.get()
+        sign = -1 if beam_y_mm > 0 else 1
+        return sign * pixel_shift * pixel_size
+
+    def _restore_pil2m_gap_offset_if_clear(self, natural_y_pos, roi_height):
+        if self._pil2m_gap_restore_y is None:
+            return False
+
+        if self._roi_overlaps_gap(natural_y_pos, roi_height, self.SAXS.detector_gap_y):
+            return False
+
+        yield from bps.mv(pil2M.motor.y, self._pil2m_gap_restore_y)
+        self._pil2m_gap_restore_y = None
+        self._pil2m_gap_y_pixel_shift = 0.0
+        self.SAXS.getPositions()
+        return True
+
+    def _avoid_pil2m_reflected_y_gap(self, y_pos, roi_height):
+        if not self._roi_overlaps_gap(y_pos, roi_height, self.SAXS.detector_gap_y):
+            return
+
+        shift_pix = self._roi_gap_clear_shift(y_pos, roi_height, self.SAXS.detector_gap_y)
+        if shift_pix == 0:
+            return
+
+        original_motor_y = pil2M.motor.y.position
+        original_beam_y = self.SAXS.direct_beam[1]
+        motor_delta = self._pil2m_motor_y_delta_for_pixel_shift(
+            shift_pix, self.SAXS.pixel_size)
+
+        if self._pil2m_gap_restore_y is None:
+            self._pil2m_gap_restore_y = original_motor_y
+
+        print(
+            "Reflected beam ROI would land in the Pilatus 2M Y gap; "
+            f"moving pil2M Y by {motor_delta:.3f} mm."
+        )
+        yield from bps.mv(pil2M.motor.y, original_motor_y + motor_delta)
+        self.SAXS.getPositions()
+        self._pil2m_gap_y_pixel_shift += self.SAXS.direct_beam[1] - original_beam_y
+
+
+    def setReflectedBeamROI(self, total_angle=0.16, technique="gisaxs", size=[48, 8],
+                            roi=pil2M.roi1,sample_z_offset_mm=0, sample_y_offset_mm=0,
+                            avoid_gaps=False):
+        """
+        Update the ROI (pil2m.roi3) for the reflected beam on the SAXS detector.
+        total_ange: float: incident angle of the alignement in degrees
+        size: tuple: size in pixels) of the ROI [width, height]
+        avoid_gaps: if True, move pil2M Y when needed so the reflected ROI avoids
+            Pilatus 2M module gaps. False preserves historical ROI-only behavior.
+        """
+
+        self.SAXS.getPositions()
+
+        # These positions are updated based on current detector position
+        x0 = self.SAXS.direct_beam[0]
+        y0 = self.SAXS.direct_beam[1]
+         
+        d = self.SAXS.distance + sample_z_offset_mm  # mm
+        pixel_size = self.SAXS.pixel_size  # mm
+
+        if technique == "gisaxs":
+            # Calculate the y position of the reflected beam
+            y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+            y_offset_pix = y_offset_mm / pixel_size
+            y_pos = int(y0 - size[1] / 2 - y_offset_pix)
+
+            restored = False
+            if avoid_gaps:
+                natural_y_pos = y_pos - self._pil2m_gap_y_pixel_shift
+                restored = yield from self._restore_pil2m_gap_offset_if_clear(
+                    natural_y_pos, size[1])
+            if restored:
+                x0 = self.SAXS.direct_beam[0]
+                y0 = self.SAXS.direct_beam[1]
+                d = self.SAXS.distance + sample_z_offset_mm
+                y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+                y_offset_pix = y_offset_mm / pixel_size
+                y_pos = int(y0 - size[1] / 2 - y_offset_pix)
+
+            if avoid_gaps:
+                yield from self._avoid_pil2m_reflected_y_gap(y_pos, size[1])
+                x0 = self.SAXS.direct_beam[0]
+                y0 = self.SAXS.direct_beam[1]
+                d = self.SAXS.distance + sample_z_offset_mm
+                y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+                y_offset_pix = y_offset_mm / pixel_size
+                y_pos = int(y0 - size[1] / 2 - y_offset_pix)
+
+            # Define the reflected beam ROI on the pilatus 2M detector
+            yield from bps.mv(roi.min_xyz.min_x, int(x0 - size[0] / 2))
+            yield from bps.mv(roi.size.x, int(size[0]))
+            yield from bps.mv(roi.min_xyz.min_y, int(y_pos))
+            yield from bps.mv(roi.size.y, int(size[1]))
+
+        elif technique == "xrr":
+            # Calculate the x position of the reflected beam
+            x_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+            x_offset_pix = x_offset_mm / pixel_size
+            x_pos = int(x0 - size[1] / 2 - x_offset_pix)
+
+            # Define the reflected beam ROI on the pilatus 2M detector
+            yield from bps.mv(roi.min_xyz.min_x, int(x_pos))
+            yield from bps.mv(roi.size.x, int(size[1]))
+            yield from bps.mv(roi.min_xyz.min_y, int(y0 - size[0] / 2))
+            yield from bps.mv(roi.size.y, int(size[0]))
+    
+        elif technique == "giwaxs": # for bdm using 900KW
+
+            # These positions are updated based on current detector position
+            x0 = self.WAXS.direct_beam[0]
+            y0 = self.WAXS.direct_beam[1]
+            d = self.WAXS.distance + sample_z_offset_mm  # mm
+            pixel_size = self.WAXS.pixel_size  # mm
+
+            # Calculate the y position of the reflected beam
+            y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+            y_offset_pix = y_offset_mm / pixel_size
+            y_pos = int(y0 - size[1] / 2 - y_offset_pix)
+
+            # Define the reflected beam ROI on the pilatus 900KW detector
+            yield from bps.mv(roi.min_xyz.min_x, int(x0 - size[0] / 2))
+            yield from bps.mv(roi.size.x, int(size[0]))
+            yield from bps.mv(roi.min_xyz.min_y, int(y_pos))
+            yield from bps.mv(roi.size.y, int(size[1]))
+        
+        elif technique == "bdm_gisaxs": # for bdm with sample tracking
+            # Calculate the y position of the reflected beam
+            y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+            y_offset_pix = y_offset_mm / pixel_size
+            # y_pos = int(y0 - size[1] / 2 - y_offset_pix)
+            sample_y_offset_pix = sample_y_offset_mm / pixel_size
+            y_pos = int(y0 - size[1] / 2 - y_offset_pix + sample_y_offset_pix)
+
+            restored = False
+            if avoid_gaps:
+                natural_y_pos = y_pos - self._pil2m_gap_y_pixel_shift
+                restored = yield from self._restore_pil2m_gap_offset_if_clear(
+                    natural_y_pos, size[1])
+            if restored:
+                x0 = self.SAXS.direct_beam[0]
+                y0 = self.SAXS.direct_beam[1]
+                d = self.SAXS.distance + sample_z_offset_mm
+                y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+                y_offset_pix = y_offset_mm / pixel_size
+                sample_y_offset_pix = sample_y_offset_mm / pixel_size
+                y_pos = int(y0 - size[1] / 2 - y_offset_pix + sample_y_offset_pix)
+
+            if avoid_gaps:
+                yield from self._avoid_pil2m_reflected_y_gap(y_pos, size[1])
+                x0 = self.SAXS.direct_beam[0]
+                y0 = self.SAXS.direct_beam[1]
+                d = self.SAXS.distance + sample_z_offset_mm
+                y_offset_mm = np.tan(np.radians(2 * total_angle)) * d
+                y_offset_pix = y_offset_mm / pixel_size
+                sample_y_offset_pix = sample_y_offset_mm / pixel_size
+                y_pos = int(y0 - size[1] / 2 - y_offset_pix + sample_y_offset_pix)
+
+            # Define the reflected beam ROI on the pilatus 2M detector
+            yield from bps.mv(roi.min_xyz.min_x, int(x0 - size[0] / 2))
+            yield from bps.mv(roi.size.x, int(size[0]))
+            yield from bps.mv(roi.min_xyz.min_y, int(y_pos))
+            yield from bps.mv(roi.size.y, int(size[1]))
+        
+        else:
+            raise ValueError("Unknown geometry fo alignement mode")
+
+    def attenuators_state(self):
+        """Summarize the inserted attenuators: per-foil material/thickness AND the
+        approximate numerical attenuation factor at the current energy.
+
+        Delegates to the ``attenuation`` device (which uses the CXRO transmission curves and
+        also mirrors this state into ``RE.md['beamline_attenuators']`` itself).  ``self.att_state``
+        is the same dict the device writes, kept here for the legacy ``update_md`` path.
+        """
+        try:
+            self.att_state = attenuation.state_md()
+        except Exception:
+            self.att_state = {"foils": {}, "attenuation_factor": None,
+                              "transmission": None, "energy_eV": None,
+                              "description": "unavailable"}
+
+    def crl_state(self):
+        for crl_le in [
+            crl.lens1,
+            crl.lens2,
+            crl.lens3,
+            crl.lens4,
+            crl.lens5,
+            crl.lens6,
+            crl.lens7,
+            crl.lens8,
+            crl.lens9,
+            crl.lens10,
+            crl.lens11,
+            crl.lens12,
+        ]:
+
+            if abs(crl_le.position) < 4:
+                self.crl_state = "micro_focusing"
+                break
+            else:
+                self.crl_state = "low_divergence"
+
+    def pressure_measurments(self):
+        if (
+            get_chamber_pressure(chamber_pressure.waxs) < 1e-02
+            and get_chamber_pressure(chamber_pressure.maxs) < 1e-02
+        ):
+            self.pressure_state = "in-vacuum"
+        elif (
+            get_chamber_pressure(chamber_pressure.waxs) > 1e-02
+            and get_chamber_pressure(chamber_pressure.maxs) < 1e-02
+        ):
+            self.pressure_state = "in-air"
+        else:
+            self.pressure_state = "low pressure on both GVs"
+
+    def update_md(self, prefix="beamline_", **md):
+        md_beamline = self.md.copy()
+
+        # RE.md[prefix + "beamsize"] = self.crl_state
+        RE.md[prefix + "sample_environment"] = self.pressure_state
+        RE.md[prefix + "attenuators"] = self.att_state
+
+        if prefix is not None:
+            md_beamline = {
+                "{:s}{:s}".format(prefix, key): value
+                for key, value in md_beamline.items()
+            }
+
+        self.md.update(md_beamline)
+
+
+# End class SMI_Beamline(Beamline)
+########################################
+
+
+def interpolate_sdds():
+    """
+    Use the redis database to interpolate the beam positions
+    """
+    pil2M.update_beam_center()
+    return pil2M.motor.z.position/1000, [pil2M.beam_center_x_px.get(), pil2M.beam_center_y_px.get()]
+
+
+class SMI_SAXS_Det(object):
+    def __init__(self, **kwargs):
+
+        # Pilatus 2M: 3 columns x 8 rows of 487 x 195 px modules.
+        # Inter-module dead areas are 7 px in X and 17 px in Y.
+        self.detector_gap_x = [[487, 494], [981, 988]]
+        self.detector_gap_y = [
+            [195, 212],
+            [407, 424],
+            [619, 636],
+            [831, 848],
+            [1043, 1060],
+            [1255, 1272],
+            [1467, 1484],
+        ]
+
+        self.pixel_size = 0.172
+        self.getPositions()
+        # self.get_beamstop()
+
+    def getPositions(self, **md):
+        """
+        Get the interpolated sample detector distance and beam position from interpolate_db_sdds
+        function defined in 01_load.py
+        """
+
+        # Interpolate the distance and direct beam position
+        self.distance, self.direct_beam = interpolate_sdds()
+        self.distance *= 1000
+
+        smi_saxs_detector.x0_pix.put(self.direct_beam[0])
+        smi_saxs_detector.y0_pix.put(self.direct_beam[1])
+        smi_saxs_detector.sdd.put(self.distance)
+
+    def get_beamstop(self):
+        """
+        Read the positions of pindiode and gisaxs beamstop to determine which beamstop is in and which pixels to mask
+        """
+
+        # ToDo: Calculate what pixels to mask for different beamstop positions
+        if saxs_bs.x_pin.position < 10 and saxs_bs.x_rod.position < 10:
+            smi_saxs_detector.bs_kind.put("rod_beamstop")
+
+            # To be implemented with the good values for y and test x position
+            x_bs = 440.79 - (saxs_bs.x_rod.position / 0.172) + (pil2m_pos.x.position / 0.172)
+            smi_saxs_detector.xbs_mask.put(x_bs)
+            y_bs = 496.07 - (saxs_bs.y_rod.position / 0.172) + (pil2m_pos.y.position / 0.172)
+            smi_saxs_detector.ybs_mask.put(10)
+
+        elif abs(saxs_bs.x_pin.position) > 50 and abs(saxs_bs.x_rod.position) < 50:
+            smi_saxs_detector.bs_kind.put("pindiode")
+
+            # To be implemented with the good values, not hard-coded
+            smi_saxs_detector.xbs_mask.put(10)
+            smi_saxs_detector.ybs_mask.put(10)
+
+        else:
+            yield from bps.mv(smi_saxs_detector.bs_kind, "None")
+
+            # To be implemented with the good values, not hard-coded
+            smi_saxs_detector.xbs_mask.put(10)
+            smi_saxs_detector.ybs_mask.put(10)
+
+    def set_beamstop(self):
+        """
+        Modify the beamstop position in order to keep the beamstop aligned with the direct beam
+        (Can be used when the sample detector distance is modified)
+        """
+
+        self.getPositions()
+        self.get_beamstop()
+
+        # ToDo: Compare beamstop position vs direct beam position and change bs position
+        pass
+
+
+
+
+class SMI_WAXS_Det(object): ### for bdm using 900KW
+    def __init__(self, **kwargs):
+
+        # ToDo: add here the position of the gap
+        self.detector_gap_x = [[830, 850], [617, 637], [405, 425], [193, 213]]
+        self.detector_gap_y = [[485, 495]]
+
+        self.pixel_size = 0.172
+        self.getPositions()
+        # self.get_beamstop()
+
+    def getPositions(self, **md):
+        """
+        Get the interpolated sample detector distance and beam position from interpolate_db_sdds
+        function defined in 01_load.py
+        """
+
+        # Interpolate the distance and direct beam position
+        # pil900KW.update_beam_center()
+        self.distance, self.direct_beam = 0.2749, [pil900KW.beam_center_x_px.get(), pil900KW.beam_center_y_px.get()]
+        self.distance *= 1000
+
+        smi_waxs_detector.x0_pix.put(self.direct_beam[0])
+        smi_waxs_detector.y0_pix.put(self.direct_beam[1])
+        smi_waxs_detector.sdd.put(self.distance)
+
+
+smi_waxs_detector = SMI_WAXS_detector(name="Pilatus900kw")
+smi_saxs_detector = SMI_SAXS_detector(name="Pilatus2M")
+
+SMI = SMI_Beamline()
+# pilatus2M = SMI_SAXS_Det()
+
+
+# from IPython import get_ipython
+# sd = get_ipython().user_ns['sd']
+# sd.baseline.extend([smi_waxs_detector, smi_saxs_detector])
